Adding api access for adding/updating/deleting an artwork

+++ src/main/java/com/artwork/online/eartwork/model/Category.java
@@ -32,7 +32,7 @@ public class Category {
     private List<Artwork> artworks;
 
 
-    public long getCategoryId() {
+    public Integer getCategoryId() {
         return categoryId;
     }
